+++ include/dcc/rx/crtp_base.hpp
@@ -705,6 +705,9 @@ private:
         (addr && addr == _addrs.consist))
       return false;
 
+    // Store packet for app:pom
+    _pom.packet = _deque.front();
+
     switch (uint32_t const cv_addr{(bytes[0uz] & 0b11u) << 8u | bytes[1uz]};
             static_cast<uint32_t>(bytes[0uz]) >> 2u & 0b11u) {
       // Reserved
@@ -719,7 +722,8 @@ private:
         if (_own_equal_packets_count < DCC_RX_MIN_CV_WRITE_PACKETS)
           ;
         // Write once
-        else if (_own_equal_packets_count == DCC_RX_MIN_CV_WRITE_PACKETS)
+        else if (_own_equal_packets_count == DCC_RX_MIN_CV_WRITE_PACKETS ||
+                 serviceMode())
           cvWrite(cv_addr, bytes[2uz]);
         // ...otherwise just verify
         else cvVerify(cv_addr, bytes[2uz]);
@@ -730,7 +734,8 @@ private:
         auto const pos{bytes[2uz] & 0b111u};
         auto const bit{static_cast<bool>(bytes[2uz] & ztl::make_mask(3u))};
         if (!(bytes[2uz] & ztl::make_mask(4u))) cvVerify(cv_addr, bit, pos);
-        else if (_own_equal_packets_count == DCC_RX_MIN_CV_WRITE_PACKETS)
+        else if (_own_equal_packets_count == DCC_RX_MIN_CV_WRITE_PACKETS ||
+                 serviceMode())
           cvWrite(cv_addr, bit, pos);
         break;
       }
@@ -938,7 +943,6 @@ private:
     else {
       _own_equal_packets_count = 1uz;
       _last_own_packet = _deque.front();
-      _pom_deque.clear();
     }
   }
 
@@ -962,13 +966,13 @@ private:
     }
   }
 
-  /// Add to pom deque
+  /// Add to PoM deque
   ///
   /// \param  value CV value
   void pom(uint8_t value) {
     if (!_ch2_data_enabled) return;
-    _pom_deque.clear();
-    _pom_deque.push_back(encode_datagram(make_datagram<Bits::_12>(0u, value)));
+    _pom.deque.clear();
+    _pom.deque.push_back(encode_datagram(make_datagram<Bits::_12>(0u, value)));
   }
 
   /// Tip-off search
@@ -1137,15 +1141,16 @@ private:
   /// Handle app:pom
   void appPom() {
     if (!_ch2_data_enabled) return;
-    // Implicitly acknowledge all CV access commands
-    else if (empty(_pom_deque))
-      impl().transmitBiDi({cbegin(acks), sizeof(acks[0uz])});
-    else {
-      auto const& datagram{_pom_deque.front()};
+    // Deque contains data for this packet
+    else if (_packet == _pom.packet && !empty(_pom.deque)) {
+      auto const& datagram{_pom.deque.front()};
       std::copy(cbegin(datagram), cend(datagram), begin(_ch2));
       impl().transmitBiDi({cbegin(_ch2), size(datagram)});
-      _pom_deque.pop_front();
+      _pom.deque.pop_front();
     }
+    // Implicitly acknowledge all CV access commands
+    else
+      impl().transmitBiDi({cbegin(acks), sizeof(acks[0uz])});
   }
 
   /// Handle app:dyn
@@ -1278,7 +1283,12 @@ private:
   ztl::inplace_deque<Datagram<datagram_size<Bits::_48>>, 1uz> _logon_deque{};
   ztl::inplace_deque<Datagram<datagram_size<Bits::_36>>, 1uz> _tos_deque{};
   ztl::inplace_deque<Datagram<datagram_size<Bits::_12>>, 2uz> _adr_deque{};
-  ztl::inplace_deque<Datagram<datagram_size<Bits::_12>>, 1uz> _pom_deque{};
+
+  // PoM
+  struct {
+    ztl::inplace_deque<Datagram<datagram_size<Bits::_12>>, 1uz> deque{};
+    Packet packet{};
+  } _pom{};
 
   Packet _packet{};          ///< Current packet
   Packet _last_own_packet{}; ///< Last packet for own address